feat: create logout and register methods, create register view page and routes

+++ app/Http/Controllers/ChirpController.php
@@ -3,12 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Models\Chirp;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
 
 class ChirpController extends Controller
 {
     
-    
+    use AuthorizesRequests;
 
     public function index()
     {
@@ -35,16 +36,12 @@ class ChirpController extends Controller
     {
         $validated = $request->validate([
             'message' => 'required|string|max:255',
-            ],[
-                'message.required' => 'Please write something to chirp!',
-                'message.max' => 'Chirps must be 255 characters or less.'
-            ]);
-
-        Chirp::create([
-            'message' => $validated['message'],
         ]);
 
-        return redirect('/')->with('success', 'Your chirp has benn posted!');
+        // Use the authenticated user
+        auth()->user()->chirps()->create($validated);
+
+        return redirect('/')->with('success', 'Your chirp has been posted!');
 
     }
 
@@ -61,31 +58,28 @@ class ChirpController extends Controller
      */
     public function edit(Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         return view('chirps.edit', compact('chirp'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         $validated = $request->validate([
             'message' => 'required|string|max:255',
-            ],[
-                'message.required' => 'Please write something to chirp!',
-                'message.max' => 'Chirps must be 255 characters or less.'
-            ]);
+        ]);
 
         $chirp->update($validated);
 
         return redirect('/')->with('success', 'Chirp updated!');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(Chirp $chirp)
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
 
         return redirect('/')->with('success', 'Chirp deleted!');
